Add customize flask-jwt-extended loader

+++ resources/item.py
@@ -1,5 +1,6 @@
+from flask_jwt_extended.view_decorators import fresh_jwt_required
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 from models.item import ItemModel
 
 
@@ -24,7 +25,7 @@ class Item(Resource):
             return item.json()
         return {'message': 'Item not found'}, 404
 
-    @jwt_required
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': "An item with name '{}' already exists.".format(name)}, 400
@@ -40,8 +41,14 @@ class Item(Resource):
 
         return item.json(), 201
 
-    @jwt_required
+    @fresh_jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {
+                'message': 'Admin privilege required'
+            }
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
